fix: keep recorder effects preview
