+++ src/components/recorder/recorder-effects-preview.tsx
@@ -1,46 +1,24 @@
 import { useState } from "react";
+import type { EqParameters } from "../../lib/dsp/biquad-eq";
 import { dbToGain } from "../../lib/music";
-import {
-  type MultibandEqState,
-  RecorderMultibandEffects,
-} from "./recorder-multiband-effects";
+import { RecorderEffectsContent } from "./recorder-effects";
 
-const INITIAL_EQ: MultibandEqState = {
+const INITIAL_EQ: EqParameters = {
+  frequency: 800,
+  gain: dbToGain(8),
+  q: 1.2,
   bypass: false,
-  bands: [
-    {
-      id: "low",
-      frequency: 120,
-      gain: dbToGain(5),
-      q: 0.8,
-      bypass: false,
-    },
-    {
-      id: "mid",
-      frequency: 850,
-      gain: dbToGain(-7),
-      q: 2.4,
-      bypass: false,
-    },
-    {
-      id: "high",
-      frequency: 4800,
-      gain: dbToGain(3.5),
-      q: 1.4,
-      bypass: false,
-    },
-  ],
 };
 
 export function RecorderEffectsPreview() {
   const [eq, setEq] = useState(INITIAL_EQ);
 
   return (
-    <div
-      data-testid="recorder-multiband-effects-panel"
-      className="w-96 rounded-lg border border-neutral-700 bg-neutral-800 px-4 py-3 shadow-2xl"
-    >
-      <RecorderMultibandEffects eq={eq} onChange={setEq} />
+    <div className="w-96 rounded-lg border border-neutral-700 bg-neutral-800 px-4 py-3 shadow-2xl">
+      <RecorderEffectsContent
+        eq={eq}
+        onChange={(update) => setEq((current) => ({ ...current, ...update }))}
+      />
     </div>
   );
 }
